test(bedrock): keep alias coverage in existing suite
